Move constants to utils directory and update import paths in files that reference it; update some styles
